requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
botocore==1.23.54
cachetools==5.2.0
cement==2.8.2
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.3
decorator==5.1.1
Django==4.1.2
flatbuffers==22.9.24
future==0.16.0
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
jmespath==0.10.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pywin32==304
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
wcwidth==0.1.9
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
